Create find_missing.py
My solution to find this question:

You are given an array of positive numbers from 1 to n, such that all numbers from 1 to n are present except one number x. You have to find x. The input array is not sorted. Look at the below array and give it a try before checking the solution.

3, 7, 1, 2, 8, 4, 5

n = 8 missing number = 6